Use UInt64 as CombineIdentifier (instead of UInt)

+++ Sources/COpenCombineHelpers/COpenCombineHelpers.cpp
@@ -9,7 +9,7 @@
 
 #include <atomic>
 
-extern "C" uintptr_t opencombine_next_combine_identifier() {
-    static std::atomic<uintptr_t> next_combine_identifier;
+extern "C" uint64_t opencombine_next_combine_identifier() {
+    static std::atomic<uint64_t> next_combine_identifier;
     return next_combine_identifier.fetch_add(1);
 }
